+++ next-cloudinary/src/components/CldVideoPlayer/index.ts
@@ -1,2 +1,2 @@
 export { default } from './CldVideoPlayer';
-export type { CldVideoPlayerPropsColors, CldVideoPlayerPropsLogo, CldVideoPlayerProps } from './CldVideoPlayer';
+export type {  CldVideoPlayerProps, CldVideoPlayerPropsColors, CldVideoPlayerPropsLogo } from './CldVideoPlayer.types';
