+++ next-cloudinary/src/index.ts
@@ -11,7 +11,9 @@ export { default as CldUploadWidget } from './components/CldUploadWidget';
 export type { CldUploadWidgetProps, CldUploadWidgetPropsChildren, CldUploadWidgetPropsOptions, CldUploadWidgetResults } from './components/CldUploadWidget';
 
 export { default as CldVideoPlayer } from './components/CldVideoPlayer';
-export type { CldVideoPlayerPropsColors, CldVideoPlayerPropsLogo, CldVideoPlayerProps } from './components/CldVideoPlayer';
+export type { CldVideoPlayerProps, CldVideoPlayerPropsColors, CldVideoPlayerPropsLogo } from './components/CldVideoPlayer';
 
 export { cloudinaryLoader } from './loaders/cloudinary-loader';
 export type { CloudinaryLoader, CloudinaryLoaderLoaderOptions, CloudinaryLoaderCldOptions } from './loaders/cloudinary-loader';
+
+export type { CloudinaryVideoPlayer } from './types/player';
