+++ next-cloudinary/src/types/player.ts
@@ -1,0 +1,3 @@
+export interface CloudinaryVideoPlayer {
+  on: Function
+}
